Added tests for broker ip settings
Broker ip settings over configuration file or system parameters are now
available. These tests cover scenarios where this value is provided or
not provided. It can be provided on a command line or in configuration file.

Default value from service configuration file was removed. This means
that value has to be provided on while running the system.

Unit tests as well as qa integration test scenarios in cucumber ware prepared
to test this functionality.

For these tests to work all other tests that use broker and connection
service had to be provided with borker ip value.

Cucumber runner was updated to support cleanup of environment variables,
as they are reused in test scenarios.

Cucumber reports in dev-tools ware updated with new cucumber tests.

Signed-off-by: Uros Mesaric Kunst <uros.mesaric-kunst@comtrade.com>